Move minified-modo from core to core/mode
