+++ build-system/babel-plugins/babel-plugin-is_minified-constant-transformer/index.js
@@ -16,7 +16,7 @@
 
 /**
  * Changes the value of IS_MINIFIED to true.
- * The above variable is in src/mode.js and src/core/minified-mode.js.
+ * The above variable is in src/mode.js and src/core/mode/minified.js.
  * @param {Object} babelTypes
  * @return {!Object}
  */
