+++ src/core/mode/minified.js
